Refactor examples folder
