+++ packages/components/psammead-episode-list/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@bbc/psammead-episode-list",
-  "version": "1.0.21",
+  "version": "1.0.22",
   "main": "dist/index.js",
   "module": "esm/index.js",
   "sideEffects": false,
